Refactor theme components and update import paths
