+++ src/pages/BeforeAfter.tsx
@@ -1,7 +1,7 @@
 
 import { motion } from 'framer-motion';
 import { Link } from 'react-router-dom';
-import Header from '../components/Header';
+import Header from '../../Header';
 import { 
   Breadcrumb, 
   BreadcrumbItem,
