+++ src/pages/Overview.tsx
@@ -2,7 +2,7 @@
 import { motion } from 'framer-motion';
 import { ChevronRight } from 'lucide-react';
 import { Link } from 'react-router-dom';
-import Header from '../components/Header';
+import Header from '../../Header';
 import OverviewCarousel from '../components/OverviewCarousel';
 
 const Overview = () => {
