+++ src/pages/USATCommerce.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { ChevronRight, ChevronUp, X } from 'lucide-react';
 import { Link } from 'react-router-dom';
-import Header from '../components/Header';
+import Header from '../../Header';
 import { Card, CardContent } from '../components/ui/card';
 const USATCommerce = () => {
   const [expandedSection, setExpandedSection] = useState<string | null>(null);
